Only require user input on non-cluster machines
When running a container on the cluster, we do not want user feedback
when deleting or not deleting temporary folders. First, the temporary
folders are fresh and removed later in a separate step. Second, we do
not want to bother the user with something like that.

We test for this by checking for the existance of the "SLURM_JOB_ID"
environment variable.

Signed-off-by: Claus Hunsen <hunsen@fim.uni-passau.de>